fix(build): move resumeData out of src/pages; lift section-sub width cap
- gatsby-plugin-page-creator tried to turn src/pages/resume/resumeData.ts
  into a page and failed the production build (no default export). Move the
  data module to src/data/ (alongside footprints-datas.ts) and update the
  resume import. `npm run build` now completes.
- Lift the .home-section-sub 62ch max-width so the subtitle sits on one line
  in the wider layout instead of wrapping to a lonely "best." orphan.
